104 My Events: The first version of the My Events is now ready. Couple of errors still needs fixing
