+++ client/src/components/MyProfile.js
@@ -3,6 +3,7 @@ import '../styles/MyProfile.css';
 import '../App.css';
 import Button from "@mui/material/Button";
 import EditDetailsModal from '../modals/EditDetailsModal';
+import EventItem from './EventItem';
 
 function ProfileItem(props) {
   // These states store the original event data
@@ -195,6 +196,8 @@ function ProfileItem(props) {
 
 function MyProfile(props) {
   const [user, setUser ] = useState(null);
+  const [events, setEvents] = useState([]);
+  const [loadingEvents, setLoadingEvents] = useState(false);
   const [selectedView, setSelectedView] = useState("Information");
 
   // These used to navigate between the My Information and My Events divs
@@ -202,6 +205,7 @@ function MyProfile(props) {
   const eventsRef = useRef(null);
   const currentRef = useRef(null);
 
+  // Used to scroll between the My Information and My Events views
   const scrollToRef = (ref) =>
   {
     if (ref.current)
@@ -223,8 +227,63 @@ function MyProfile(props) {
     }
   };
 
+  const fetchUsersEvents = async () =>
+  {
+    fetch('/api/get/events/for/' + props.currentUser.id, {
+      method: "GET"
+    })
+    .then(response => response.json())
+    .then(data =>
+      {
+        setEvents(data.events);
+      });
+  }
+
   useEffect(() => {
+    // Fetches the users personal information
     fetchUserData();
+
+    // WE DON'T CURRENTLY HAVE THE ABILITY TO JOIN EVENTS AND FOR THAT REASON YOU WON'T SEE ANY EVENTS IN
+    // THE MY EVENTS VIEW. IF YOU WANT TO SEE WHAT THE MY EVENTS LIST WOULD LOOK LIKE, COMMENT THE 
+    // 'fetchUsersEvents'- METHOD AND UNCOMMENT THE LIST CREATION BELOW. THIS FUNCTIONALITY WILL BE
+    // DELETED ONCE WE HAVE THE ABILITY TO JOIN EVENTS AND THE "REAL" EVENTS LIST CAN BE SHOWN HERE
+
+    // Fetches the events the user is currently partisipating in
+    // fetchUsersEvents();
+
+    // Create events for testing
+    
+    setEvents([
+      {
+        title: "TechSynergy Summit",
+        creator: "Emily Thompson",
+        time: "26.2.2024 12:00",
+        location: "LUT University",
+        attendees: 41,
+        attending: true,
+        description: `The TechSynergy Summit is a premier corporate event designed to bring together industry leaders, visionaries, and innovators in the ever-evolving landscape of technology. This exclusive summit serves as a dynamic platform or collaboration, knowledge exchange, and networking. Attendees can expect insightful keynote presentations, interactive panel discussions, and hands-on workshops that explore the intersection of cutting-edge technologies, fostering an environment where ideas converge, and innovation thrives. Join us at TechSynergy Summit to be at the forefront of the technological revolution and cultivate meaningful connections that propel your organization into the future.`,
+      },
+      {
+        title: "FutureTech Showcase",
+        creator: "Liam Patel",
+        time: "2.4.2024 14:00",
+        location: "LUT University",
+        attendees: 24,
+        attending: true,
+        description: `Step into the future with FutureTech Showcase, where cutting-edge innovations and breakthrough technologies converge. Explore the latest advancements in robotics, artificial intelligence, and beyond. Immerse yourself in a curated exhibition that unveils tomorrow's tech landscape today.`,
+      },
+      {
+        title: "Digital Nexus Symposium",
+        creator: "Sophia Mitchell",
+        time: "18.7.2024 10:00",
+        location: "LUT University",
+        attendees: 98,
+        attending: true,
+        description: `Join thought leaders and industry experts at the Digital Nexus Symposium, a dynamic gathering that explores the interconnected world of digital technologies. Engage in insightful discussions on the impact of AI, data analytics, and cyber-physical systems. Discover the converging points shaping our digital future at this symposium of ideas and collaboration.`,
+      },
+    ]);
+    
+
   }, []);
 
   return (
@@ -261,7 +320,41 @@ function MyProfile(props) {
         </div>
         {/* This is shwon when the user has selected "My Events" from the top menu. */}
         <div className='MyEventsArea' ref={eventsRef}>
-          
+          {loadingEvents ?
+            (
+              <p className='HintParagraphBig'>Loading...</p>
+            )
+            :
+            (
+              <div>
+                {events.length === 0 ?
+                  (
+                    <p className='HintParagraphBig' style={{margin: "30px 0 0 0", fontStyle: "italic"}}>You are not partisipating in any events</p>
+                  ) 
+                  :
+                  (
+                    <div style={{width: "100%", display: "flex",flexDirection: "column", alignItems: "center", overflow: "scroll", maxHeight: "65vh"}}>
+                      {events.map((event, index) => 
+                      (
+                        <EventItem
+                        // The admin abilities are always disabled meaning that the admin can only edit the events from the Home view.
+                        admin={false}
+                        title={event.title}
+                        creator={event.creator}
+                        time={event.time}
+                        location={event.location}
+                        attendees={event.attendees}
+                        attending={event.attending}
+                        description={event.description}
+                        key={index}
+                        />
+                      ))}
+                    </div>
+                  ) 
+                }
+              </div>
+            )
+          }
         </div>
       </div>
     </div>
