+++ client/src/styles/MyProfile.css
@@ -89,7 +89,12 @@ img
 {
   width: 100%;
   min-width: 100%;
-  padding: 15px 0 15px 0;
+  height: 100%;
 
-  background-color: antiquewhite;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  
+  padding: 15px 0 15px 0;
+  overflow: scroll;
 }
